Add `GlobalContext` type
This can be shared between different services.
It is supposed to be thread safe and use only
async-aware syncing methods.

Currently this is used to hold run config & db connection.
This is not kind great, because there will be only a single db
connection, but it should not be that complitaed to exchange it later
for some db proxy with connection pool support.