[server,cli,proto,docs] Add first implementation of AddFile command (#41)
* Update notes on db design

* Create & run migration for `local_files` table

This is the table responsible for storing local files that the deamon
exposes for syncing.

* Create & run migration for creation of `synced_files` table

It contains data to track what file & on what host we track.

* Add `AddFile` rpc to ClientApi service

* Add client implementation of `AddFile` command

I don't like the fact that `Command.handle_add_file` takes
`self` by reference instead of moving it like the rest of the commands.
This was done this way, because I need to pass down the file_path and
this paritally moves self & I can not later call method that requires
moving.

I think that I should refactor all the methods - move the
implementations under separate type / namespace and do not take "self"
at all, since I do not need it.

* Add dependency on sha1_smol

Decided to go with `sha1_smol` and not `sha1` since I do not want whole
dependency chain.

* Implement function to has a file

I've decided to stream the file into hashing instance instead
of reading whole file at once, to cap potential memory usage.

* Implement model & method for saving local file info to db

Currently this is happy-path-only implementation. No errors are handled.
This needs to be done.

* [server] First implementation of add_file server method

WARNING!

This method currently has a bug -> e.g. when client requests to add
`Cargo.toml` & the path is resolved to absolute path on server it might
actually point to `.../dsync-server/Cargo.toml` instead of expected
`.../dsync-cli/Cargo.toml`.

Path must be resolved on client side!

* Fix incorrect absolute path resolution

When path was canonicalized on server it would resolve to a file given
cwd of current server process.

Now the path is canonicalized on the client side.